requirements.txt
pyyaml
pytest
jsonschema